types.h : TPort : become u16 instead of u32, because a port is between 0 and 65535

+++ src/types.h
@@ -31,8 +31,8 @@ typedef s32 TSocket;
 // AF_INET : use IP
 // AF_VSOCK : use vsock
 
-// unsigned int
-typedef u32 TPort;
+// unsigned short
+typedef u16 TPort;
 
 // char
 typedef s8 TPower; // empirical observed values with int : [-123,20]
